app/requirements.txt
flask
pycryptodome
redis